move launchabelActivity detection to Run task, add memoization for deviceId detection
